Fix std::max type mismatch compilation errors on macOS in remaining RayTracing benchmarks

+++ cpp_src/benchmarks/RayAnyHitBench.cpp
@@ -228,7 +228,7 @@ void RayAnyHitBench::buildAS() {
   size_t triScratch = createTLAS(topTriGeom, triangleTlas, triangleTlasBuffer);
 
   // Scratch buffer
-  size_t scratchSize = std::max(triSizes.buildScratchSize, triScratch);
+  size_t scratchSize = std::max(static_cast<size_t>(triSizes.buildScratchSize), triScratch);
   scratchBuffer = context->createBuffer(scratchSize);
   VkDeviceAddress sAddr = vContext->getBufferDeviceAddress(scratchBuffer);
